Renames reportMaker file
