+++ src/domains/FeedbackPortal/ReportSummary/ReportSummary.test.tsx
@@ -6,7 +6,7 @@ import faker from 'faker';
 import { AxiosResponse } from 'axios';
 
 import ReportEndpointContext from '../Contexts/ReportEndpointContext';
-import { makeFeedbackReport, makeBugReport } from '../reportMaker.mocks';
+import { makeFeedbackReport, makeBugReport } from '../reportMaker.mock';
 import ReportSummary from './ReportSummary';
 import * as API from '../api/api'; // babel issues ref: https://stackoverflow.com/questions/53162001/typeerror-during-jests-spyon-cannot-set-property-getrequest-of-object-which
 import { FeedbackForm, BugReportForm } from '../types';
